# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SurveyBridge API",
    "version": "1.0.0",
    "description": "Survey sampling orchestration platform. Integrates SurveyMonkey (survey hosting) with PureSpectrum Fusion (panel sampling). The platform tracks completion events and forwards them to client webhooks — it never stores survey answers."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/webhook/fusion": {
      "post": {
        "tags": ["Responses"],
        "summary": "Receive Fusion event webhook",
        "description": "Server-to-server webhook called by PureSpectrum Fusion when a respondent completes or is screened out. Validated via HMAC-SHA256 signature in X-Fusion-Signature header. No JWT required.",
        "operationId": "fusionWebhook",
        "security": [],
        "parameters": [
          { "name": "X-Fusion-Signature", "in": "header", "required": false, "schema": { "type": "string" }, "description": "HMAC-SHA256 hex signature of the request body" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FusionEventPayload" } } }
        },
        "responses": {
          "202": { "description": "Accepted — event queued for async processing" },
          "403": { "description": "Invalid or missing signature" }
        }
      }
    },
    "/api/v1/surveys/{surveyId}/responses": {
      "get": {
        "tags": ["Responses"],
        "summary": "List events for a survey",
        "operationId": "listResponses",
        "parameters": [
          { "$ref": "#/components/parameters/SurveyIdPath" },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "Filter from (ISO 8601)" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "Filter to (ISO 8601)" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Paginated list of events", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PagedResponseDto" } } } }
        }
      }
    },
    "/api/v1/surveys/{surveyId}/responses/export": {
      "get": {
        "tags": ["Responses"],
        "summary": "Export events as CSV or Excel",
        "operationId": "exportResponses",
        "parameters": [
          { "$ref": "#/components/parameters/SurveyIdPath" },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["csv", "excel"], "default": "csv" } }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/csv": { "schema": { "type": "string", "format": "binary" } },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": { "schema": { "type": "string", "format": "binary" } }
            }
          }
        }
      }
    },
    "/complete": {
      "get": {
        "tags": ["Redirects"],
        "summary": "Completion redirect from SurveyMonkey",
        "description": "SurveyMonkey redirects the respondent's browser here after survey completion. No auth required. The authoritative event comes via the Fusion server-side webhook.",
        "operationId": "completionRedirect",
        "security": [],
        "parameters": [
          { "name": "UID", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Respondent ID assigned by Fusion" }
        ],
        "responses": {
          "200": { "description": "Acknowledged" }
        }
      }
    },
    "/terminate": {
      "get": {
        "tags": ["Redirects"],
        "summary": "Screenout redirect from SurveyMonkey",
        "description": "SurveyMonkey redirects the respondent's browser here after a screenout. No auth required.",
        "operationId": "screenoutRedirect",
        "security": [],
        "parameters": [
          { "name": "UID", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Respondent ID assigned by Fusion" }
        ],
        "responses": {
          "200": { "description": "Acknowledged" }
        }
      }
    },
...
    "/api/v1/dashboard/surveys/{surveyId}/timeseries": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get response time series for a survey",
        "operationId": "timeSeries",
        "parameters": [
          { "$ref": "#/components/parameters/SurveyIdPath" },
          { "name": "from", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "granularity", "in": "query", "schema": { "type": "string", "enum": ["day", "week", "month"], "default": "day" } }
        ],
        "responses": {
          "200": { "description": "Time series data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TimeSeriesResponseDto" } } } }
        }
      }
    },
...
    "/api/v1/notifications/{id}/read": {
      "put": {
        "tags": ["Notifications"],
        "summary": "Mark notification as read",
        "operationId": "markRead",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "204": { "description": "Marked as read" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/v1/admin/clients": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all clients",
        "operationId": "listClients",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Paginated client list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PagedAdminClientDto" } } } }
        }
      }
    },
    "/api/v1/admin/clients/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get client by ID",
        "operationId": "getClient",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Client details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminClientDto" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/admin/clients/{id}/quota": {
      "put": {
        "tags": ["Admin"],
        "summary": "Update client monthly quota",
        "operationId": "updateQuota",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateQuotaRequest" } } }
        },
        "responses": {
          "204": { "description": "Updated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak-issued JWT. Obtain via Keycloak token endpoint."
      }
    },
    "parameters": {
...
      "SurveyIdPath": {
        "name": "surveyId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Survey UUID"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
      },
...
    },
    "schemas": {
...
      "EventType": {
        "type": "string",
        "enum": ["COMPLETE", "SCREENOUT"]
      },
...
      "FusionEventPayload": {
        "type": "object",
        "required": ["event", "respondent_id", "fusion_survey_id"],
        "properties": {
          "event": { "type": "string", "enum": ["complete", "screenout"], "example": "complete" },
          "respondent_id": { "type": "string", "example": "R12345" },
          "fusion_survey_id": { "type": "string", "example": "FS_7788" },
          "cpi": { "type": "number", "format": "decimal", "example": 3.2 }
        }
      },
      "ResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "surveyId": { "type": "string", "format": "uuid" },
          "respondentId": { "type": "string", "example": "R12345" },
          "fusionSurveyId": { "type": "string", "example": "FS_7788" },
          "eventType": { "$ref": "#/components/schemas/EventType" },
          "cpi": { "type": "number", "nullable": true, "example": 3.2 },
          "occurredAt": { "type": "string", "format": "date-time" }
        }
      },
      "PagedResponseDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/ResponseDto" } },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer" },
          "page": { "type": "integer" },
          "size": { "type": "integer" }
        }
      },
...
      "TimeSeriesPoint": {
        "type": "object",
        "properties": {
          "period": { "type": "string", "example": "2025-04-20" },
          "responses": { "type": "integer", "format": "int64", "description": "Total events (completes + screenouts)" },
          "completions": { "type": "integer", "format": "int64", "description": "Complete events only" }
        }
      },
      "TimeSeriesResponseDto": {
        "type": "object",
        "properties": {
          "surveyId": { "type": "string", "format": "uuid" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/TimeSeriesPoint" } }
        }
      },
...
      "AdminClientDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "contactEmail": { "type": "string", "format": "email" },
          "plan": { "type": "string" },
          "monthlyResponseQuota": { "type": "integer" },
          "usedResponseCount": { "type": "integer" },
          "active": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PagedAdminClientDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/AdminClientDto" } },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer" },
          "page": { "type": "integer" },
          "size": { "type": "integer" }
        }
      },
      "UpdateQuotaRequest": {
        "type": "object",
        "required": ["monthlyResponseQuota"],
        "properties": {
          "monthlyResponseQuota": { "type": "integer", "minimum": 0, "example": 5000 }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      }
    }
  }
}